requirements.txt
pandas==2.1.4
polars==0.20.2
requests==2.31.0
jupyter==1.0.0
pyarrow==14.0.1
python-dotenv==1.0.0
pydantic==2.5.0
yfinance==0.2.33
matplotlib==3.8.0
